+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/controller/NamespaceController.java
@@ -101,6 +101,11 @@ public class NamespaceController {
     return appNamespaceService.findPublicAppNamespaces();
   }
 
+  @GetMapping("/appnamespaces/public/names")
+  public List<String> findPublicAppNamespaceNames() {
+    return appNamespaceService.findPublicAppNamespaceNames();
+  }
+
   @GetMapping("/apps/{appId}/envs/{env}/clusters/{clusterName}/namespaces")
   public List<NamespaceBO> findNamespaces(@PathVariable String appId, @PathVariable String env,
       @PathVariable String clusterName) {