+++ apollo-portal/src/main/java/com/ctrip/framework/apollo/portal/service/AppNamespaceService.java
@@ -70,6 +70,10 @@ public class AppNamespaceService {
     return appNamespaceRepository.findByIsPublicTrue();
   }
 
+  public List<String> findPublicAppNamespaceNames() {
+    return appNamespaceRepository.findNamesByIsPublicTrue();
+  }
+
   public AppNamespace findPublicAppNamespace(String namespaceName) {
     List<AppNamespace> appNamespaces =
         appNamespaceRepository.findByNameAndIsPublic(namespaceName, true);